Created new component/page MovieList
